third_party/WebKit/LayoutTests/fast/dom/shadow/part-pseudo-element-expected.txt

Test whether ::part matches elements whose part attributes have the same as ::part's argument in shadow trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS backgroundColorOf('host/target') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host1/host2/target') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host/target') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host/target') is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
